+++ QuickDraw.sln
@@ -5,16 +5,22 @@ VisualStudioVersion = 17.6.33712.159
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuickDraw", "QuickDrawWindows\QuickDraw.csproj", "{CBC43D79-6F32-480A-BC3A-54383C19E4DD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuickDraw.Core", "QuickDraw.Core\QuickDraw.Core.csproj", "{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|ARM64 = Debug|ARM64
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|ARM64 = Release|ARM64
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|Any CPU.Build.0 = Debug|x64
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|ARM64.ActiveCfg = Debug|ARM64
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|ARM64.Build.0 = Debug|ARM64
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|ARM64.Deploy.0 = Debug|ARM64
@@ -24,6 +30,8 @@ Global
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|x86.ActiveCfg = Debug|x86
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|x86.Build.0 = Debug|x86
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Debug|x86.Deploy.0 = Debug|x86
+		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|Any CPU.ActiveCfg = Release|x64
+		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|Any CPU.Build.0 = Release|x64
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|ARM64.ActiveCfg = Release|ARM64
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|ARM64.Build.0 = Release|ARM64
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|ARM64.Deploy.0 = Release|ARM64
@@ -33,6 +41,22 @@ Global
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|x86.ActiveCfg = Release|x86
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|x86.Build.0 = Release|x86
 		{CBC43D79-6F32-480A-BC3A-54383C19E4DD}.Release|x86.Deploy.0 = Release|x86
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|x64.Build.0 = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Debug|x86.Build.0 = Debug|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|ARM64.Build.0 = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|x64.ActiveCfg = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|x64.Build.0 = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|x86.ActiveCfg = Release|Any CPU
+		{DC397900-3CE2-426A-B15B-A7CA3C2EEE2C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
