+++ QuickDrawWindows/Activation/DefaultActivationHandler.cs
@@ -1,0 +1,26 @@
+﻿using Microsoft.UI.Xaml;
+using QuickDraw.Contracts.Services;
+using QuickDraw.ViewModels;
+using System.Threading.Tasks;
+
+namespace QuickDraw.Activation;
+
+public class DefaultActivationHandler : ActivationHandler<LaunchActivatedEventArgs>
+{
+    private readonly INavigationService _navigationService;
+
+    public DefaultActivationHandler(INavigationService navigationService) => _navigationService = navigationService;
+
+    protected override bool CanHandleInternal(LaunchActivatedEventArgs args)
+    {
+        // None of the ActivationHandlers has handled the activation.
+        return _navigationService.Frame?.Content == null;
+    }
+
+    protected async override Task HandleInternalAsync(LaunchActivatedEventArgs args)
+    {
+        _navigationService.NavigateTo(typeof(MainViewModel).FullName!, args.Arguments);
+
+        await Task.CompletedTask;
+    }
+}
