+++ QuickDrawWindows/App.xaml
@@ -3,7 +3,6 @@
     x:Class="QuickDraw.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:QuickDraw"
     RequestedTheme="Dark">
     <Application.Resources>
         <ResourceDictionary>
