+++ QuickDrawWindows/Contracts/Services/IActivationService.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace QuickDraw.Contracts.Services;
+
+public interface IActivationService
+{
+    Task ActivateAsync(object activationArgs);
+}
