+++ QuickDrawWindows/Contracts/Services/INavigationService.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Navigation;
+
+namespace QuickDraw.Contracts.Services;
+
+public interface INavigationService
+{
+    event NavigatedEventHandler Navigated;
+
+    bool CanGoBack
+    {
+        get;
+    }
+
+    Frame? Frame
+    {
+        get; set;
+    }
+
+    bool NavigateTo(string pageKey, object? parameter = null, bool clearNavigation = false);
+
+    bool GoBack();
+
+    void SetListDataItemForNextConnectedAnimation(object item);
+}
