+++ QuickDrawWindows/Contracts/Services/IPageService.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace QuickDraw.Contracts.Services;
+
+public interface IPageService
+{
+    Type GetPageType(string key);
+}
