+++ QuickDrawWindows/Contracts/Services/ISettingsService.cs
@@ -1,0 +1,13 @@
+﻿using QuickDraw.Core.Models;
+using System.Threading.Tasks;
+
+namespace QuickDraw.Contracts.Services;
+
+public interface ISettingsService
+{
+    public Settings? Settings { get; }
+
+    public Task InitializeAsync();
+    public Task ReadSettings();
+    public Task WriteSettings();
+}
