+++ QuickDrawWindows/Contracts/ViewModels/INavigationAware.cs
@@ -1,0 +1,9 @@
+﻿namespace QuickDraw.Contracts.ViewModels;
+
+public interface INavigationAware
+{
+    void OnNavigatedTo(object parameter);
+
+    void OnNavigatedFrom();
+}
+
