+++ QuickDrawWindows/Services/ActivationService.cs
@@ -1,0 +1,54 @@
+﻿using Microsoft.UI.Xaml;
+using QuickDraw.Activation;
+using QuickDraw.Contracts.Services;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace QuickDraw.Services;
+
+public class ActivationService(ActivationHandler<LaunchActivatedEventArgs> defaultHandler, IEnumerable<IActivationHandler> activationHandlers, ISettingsService settingsService) : IActivationService
+{
+    private readonly ActivationHandler<LaunchActivatedEventArgs> _defaultHandler = defaultHandler;
+    private readonly IEnumerable<IActivationHandler> _activationHandlers = activationHandlers;
+    private readonly ISettingsService _settingsService = settingsService;
+
+    public async Task ActivateAsync(object activationArgs)
+    {
+        await InitializeAsync();
+
+        App.Window.Activate();
+
+        await HandleActivationAsync(activationArgs);
+
+        await StartupAsync();
+    }
+
+    private async Task HandleActivationAsync(object activationArgs)
+    {
+        var activationHandler = _activationHandlers.FirstOrDefault(h => h.CanHandle(activationArgs));
+
+        if (activationHandler != null)
+        {
+            await activationHandler.HandleAsync(activationArgs);
+        }
+
+        if (_defaultHandler.CanHandle(activationArgs))
+        {
+            await _defaultHandler.HandleAsync(activationArgs);
+        }
+    }
+
+    private async Task InitializeAsync()
+    {
+        await _settingsService.InitializeAsync().ConfigureAwait(false);
+        await Task.CompletedTask;
+    }
+
+    private async Task StartupAsync()
+    {
+        await Task.CompletedTask;
+    }
+}
