+++ QuickDrawWindows/Services/PageService.cs
@@ -1,0 +1,59 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+using System;
+using System.Collections.Generic;
+using Microsoft.UI.Xaml.Controls;
+
+using QuickDraw.Contracts.Services;
+using System.Linq;
+using QuickDraw.Views;
+using QuickDraw.ViewModels;
+
+namespace QuickDraw.Services;
+
+public class PageService : IPageService
+{
+    private readonly Dictionary<string, Type> _pages = [];
+
+    public PageService()
+    {
+        // Configure Pages and their View Models
+        Configure<MainViewModel, MainPage>();
+    }
+
+    public Type GetPageType(string key)
+    {
+        Type? pageType;
+        lock (_pages)
+        {
+            if (!_pages.TryGetValue(key, out pageType))
+            {
+                throw new ArgumentException($"Page not found: {key}. Did you forget to call PageService.Configure?");
+            }
+        }
+
+        return pageType;
+    }
+
+    private void Configure<VM, V>()
+    where VM : ObservableObject
+    where V : Page
+    {
+        lock (_pages)
+        {
+            var key = typeof(VM).FullName!;
+            if (_pages.ContainsKey(key))
+            {
+                throw new ArgumentException($"The key {key} is already configured in PageService");
+            }
+
+            var type = typeof(V);
+            if (_pages.ContainsValue(type))
+            {
+                throw new ArgumentException($"This type is already configured with key {_pages.First(p => p.Value == type).Key}");
+            }
+
+            _pages.Add(key, type);
+        }
+    }
+}
