+++ QuickDrawWindows/Themes/Generic.xaml
@@ -1,12 +1,13 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:QuickDraw">
+    xmlns:local="using:QuickDraw"
+    xmlns:views="using:QuickDraw.Views">
 
-    <Style TargetType="local:SlideTitleBar" >
+    <Style TargetType="views:SlideTitleBar" >
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="local:SlideTitleBar">
+                <ControlTemplate TargetType="views:SlideTitleBar">
                     <Grid Width="Auto" Height="48" x:Name="LayoutRoot" >
                         <Grid.Resources>
 
@@ -79,10 +80,10 @@
         </Setter>
     </Style>
 
-    <Style TargetType="local:MainTitleBar" >
+    <Style TargetType="views:MainTitleBar" >
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="local:MainTitleBar">
+                <ControlTemplate TargetType="views:MainTitleBar">
                     <Grid Width="Auto" Height="32" x:Name="LayoutRoot" >
 
                         <VisualStateManager.VisualStateGroups>
