+++ QuickDrawWindows/Utilities/Binding.cs
@@ -4,13 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace QuickDraw.Utilities
+namespace QuickDraw.Utilities;
+
+public class Binding
 {
-    public class Binding
+    public static bool InvertBool(bool value)
     {
-        public static bool InvertBool(bool value)
-        {
-            return !value;
-        }
+        return !value;
     }
 }
