+++ QuickDrawWindows/Utilities/FolderDialog.cs
@@ -1,0 +1,312 @@
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+namespace QuickDraw.Utilities;
+
+// Disable warning CS0108: 'x' hides inherited member 'y'. Use the new keyword if hiding was intended.
+#pragma warning disable 0108
+
+internal static class IIDGuid
+{
+    internal const string IModalWindow = "b4db1657-70d7-485e-8e3e-6fcb5a5c1802";
+    internal const string IFileDialog = "42f85136-db7e-439c-85f1-e4075d135fc8";
+    internal const string IFileOpenDialog = "d57c7288-d4ad-4768-be02-9d969532d960";
+    internal const string IShellItem = "43826d1e-e718-42ee-bc55-a1e261c37bfe";
+    internal const string IShellItemArray = "B63EA76D-1F85-456F-A19C-48159EFA858B";
+}
+
+internal static class CLSIDGuid
+{
+    internal const string FileOpenDialog = "DC1C5A9C-E88A-4dde-A5A1-60F82A20AEF7";
+    internal const string ShellItem = "9ac9fbe1-e0a2-4ad6-b4ee-e212013ea917";
+}
+
+[Flags]
+internal enum FileOpenDialogOptions : uint
+{
+    NoChangeDir = 0x00000008,
+    PickFolders = 0x00000020,
+    AllowMultiSelect = 0x00000200,
+    PathMustExist = 0x00000800,
+}
+
+public enum SIGDN : uint
+{
+    NORMALDISPLAY = 0,
+    PARENTRELATIVEPARSING = 0x80018001,
+    PARENTRELATIVEFORADDRESSBAR = 0x8001c001,
+    DESKTOPABSOLUTEPARSING = 0x80028000,
+    PARENTRELATIVEEDITING = 0x80031001,
+    DESKTOPABSOLUTEEDITING = 0x8004c000,
+    FILESYSPATH = 0x80058000,
+    URL = 0x80068000
+}
+
+[ComImport]
+[Guid(IIDGuid.IShellItem)]
+[InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+public interface IShellItem
+{
+    void BindToHandler(nint pbc,
+        [MarshalAs(UnmanagedType.LPStruct)] Guid bhid,
+        [MarshalAs(UnmanagedType.LPStruct)] Guid riid,
+        out nint ppv);
+
+    void GetParent(out IShellItem ppsi);
+
+    void GetDisplayName(SIGDN sigdnName, out nint ppszName);
+
+    void GetAttributes(uint sfgaoMask, out uint psfgaoAttribs);
+
+    void Compare(IShellItem psi, uint hint, out int piOrder);
+};
+
+[ComImport,
+Guid(IIDGuid.IShellItemArray),
+InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+interface IShellItemArray
+{
+    // Not supported: IBindCtx
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void BindToHandler([In, MarshalAs(UnmanagedType.Interface)] nint pbc, [In] ref Guid rbhid,
+                 [In] ref Guid riid, out nint ppvOut);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetPropertyStore([In] int Flags, [In] ref Guid riid, out nint ppv);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetPropertyDescriptionList([In] ref PropertyKey keyType, [In] ref Guid riid, out nint ppv);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetAttributes([In] ShellItemArrayAttributeFlags dwAttribFlags, [In] uint sfgaoMask, out uint psfgaoAttribs);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetCount(out uint pdwNumItems);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetItemAt([In] uint dwIndex, [MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    // Not supported: IEnumShellItems (will use GetCount and GetItemAt instead)
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void EnumItems([MarshalAs(UnmanagedType.Interface)] out nint ppenumShellItems);
+}
+
+
+
+// These two are here for methods we don't use (but need to keep the function order), so they are empty
+struct COMDLG_FILTERSPEC { }
+
+internal interface IFileDialogEvents { }
+
+internal enum FileDialogAddPosition : uint
+{
+
+}
+
+public struct PropertyKey
+{
+
+}
+
+public enum ShellItemArrayAttributeFlags
+{
+
+}
+
+// End Empty Definitions
+
+
+[ComImport(),
+Guid(IIDGuid.IModalWindow),
+InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+internal interface IModalWindow
+{
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime),
+    PreserveSig]
+    int Show([In] nint parent);
+}
+
+[ComImport(),
+Guid(IIDGuid.IFileDialog),
+InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+internal interface IFileDialog : IModalWindow
+{
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime),
+    PreserveSig]
+    int Show([In] nint parent);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileTypes([In] uint cFileTypes, [In] COMDLG_FILTERSPEC[] rgFilterSpec);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileTypeIndex([In] uint iFileType);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetFileTypeIndex(out uint piFileType);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void Advise([In, MarshalAs(UnmanagedType.Interface)] IFileDialogEvents pfde, out uint pdwCookie);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void Unadvise([In] uint dwCookie);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetOptions([In] FileOpenDialogOptions fos);
+
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetOptions(out FileOpenDialogOptions pfos);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetDefaultFolder([In, MarshalAs(UnmanagedType.Interface)] IShellItem psi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFolder([In, MarshalAs(UnmanagedType.Interface)] IShellItem psi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetFolder([MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetCurrentSelection([MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileName([In, MarshalAs(UnmanagedType.LPWStr)] string pszName);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetFileName([MarshalAs(UnmanagedType.LPWStr)] out string pszName);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetTitle([In, MarshalAs(UnmanagedType.LPWStr)] string pszTitle);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetOkButtonLabel([In, MarshalAs(UnmanagedType.LPWStr)] string pszText);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileNameLabel([In, MarshalAs(UnmanagedType.LPWStr)] string pszLabel);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetResult([MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void AddPlace([In, MarshalAs(UnmanagedType.Interface)] IShellItem psi, FileDialogAddPosition fdap);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetDefaultExtension([In, MarshalAs(UnmanagedType.LPWStr)] string pszDefaultExtension);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void Close([MarshalAs(UnmanagedType.Error)] int hr);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetClientGuid([In] ref Guid guid);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void ClearClientData();
+
+    // Not supported:  IShellItemFilter is not defined, converting to IntPtr
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFilter([MarshalAs(UnmanagedType.Interface)] nint pFilter);
+}
+
+[ComImport(),
+Guid(IIDGuid.IFileOpenDialog),
+InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+internal interface IFileOpenDialog : IFileDialog
+{
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime),
+    PreserveSig]
+    int Show([In] nint parent);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileTypes([In] uint cFileTypes, [In] COMDLG_FILTERSPEC[] rgFilterSpec);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileTypeIndex([In] uint iFileType);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetFileTypeIndex(out uint piFileType);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void Advise([In, MarshalAs(UnmanagedType.Interface)] IFileDialogEvents pfde, out uint pdwCookie);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void Unadvise([In] uint dwCookie);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetOptions([In] FileOpenDialogOptions fos);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetOptions(out FileOpenDialogOptions pfos);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetDefaultFolder([In, MarshalAs(UnmanagedType.Interface)] IShellItem psi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFolder([In, MarshalAs(UnmanagedType.Interface)] IShellItem psi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetFolder([MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetCurrentSelection([MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileName([In, MarshalAs(UnmanagedType.LPWStr)] string pszName);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetFileName([MarshalAs(UnmanagedType.LPWStr)] out string pszName);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetTitle([In, MarshalAs(UnmanagedType.LPWStr)] string pszTitle);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetOkButtonLabel([In, MarshalAs(UnmanagedType.LPWStr)] string pszText);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFileNameLabel([In, MarshalAs(UnmanagedType.LPWStr)] string pszLabel);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetResult([MarshalAs(UnmanagedType.Interface)] out IShellItem ppsi);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void AddPlace([In, MarshalAs(UnmanagedType.Interface)] IShellItem psi, FileDialogAddPosition fdap);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetDefaultExtension([In, MarshalAs(UnmanagedType.LPWStr)] string pszDefaultExtension);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void Close([MarshalAs(UnmanagedType.Error)] int hr);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetClientGuid([In] ref Guid guid);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void ClearClientData();
+
+    // Not supported:  IShellItemFilter is not defined, converting to IntPtr
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void SetFilter([MarshalAs(UnmanagedType.Interface)] nint pFilter);
+
+    // Defined by IFileOpenDialog
+    // ---------------------------------------------------------------------------------
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetResults([MarshalAs(UnmanagedType.Interface)] out IShellItemArray ppenum);
+
+    [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+    void GetSelectedItems([MarshalAs(UnmanagedType.Interface)] out IShellItemArray ppsai);
+}
+
+[ComImport,
+ClassInterface(ClassInterfaceType.None),
+TypeLibType(TypeLibTypeFlags.FCanCreate),
+Guid(CLSIDGuid.FileOpenDialog)]
+internal class FileOpenDialogRCW
+{
+}
+
+[ComImport,
+Guid(IIDGuid.IFileOpenDialog),
+CoClass(typeof(FileOpenDialogRCW))]
+internal interface NativeFileOpenDialog : IFileOpenDialog
+{
+}
