+++ QuickDrawWindows/ViewModels/MainViewModel.cs
@@ -1,0 +1,8 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace QuickDraw.ViewModels;
+
+public partial class MainViewModel : ObservableRecipient
+{
+
+}
