+++ QuickDrawWindows/Views/ImageFolderListViewControl.xaml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <UserControl
-    x:Class="QuickDraw.MFImageFolderListView"
+    x:Class="QuickDraw.Views.ImageFolderListViewControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:QuickDraw"
+    xmlns:local="using:QuickDraw.Views"
+    xmlns:models="using:QuickDraw.Core.Models"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
@@ -38,8 +39,8 @@
 
         <ListView x:Name="ImageFolderListView" PointerMoved="ImageFolderListView_PointerMoved" DragItemsStarting="ImageFolderListView_DragItemsStarting" DragItemsCompleted="ImageFolderListView_DragItemsCompleted" CanDragItems="True" ItemsSource="{x:Bind ImageFolderCollection}" Background="{StaticResource CardBackgroundFillColorDefaultBrush}" CornerRadius="6" IsItemClickEnabled="False" CanReorderItems="True" AllowDrop="True" SelectionMode="Multiple" Grid.Row="1" BorderThickness="1" BorderBrush="{StaticResource CardStrokeColorDefaultBrush}" >
             <ListView.ItemTemplate>
-                <DataTemplate x:DataType="local:Models.MFImageFolder">
-                    <local:MFImageFolderView Folder="{x:Bind}" SizeChanged="MFImageFolderControl_SizeChanged" DesiredPathColumnWidth="{Binding ElementName=Control, Path=DesiredPathColumnWidth}" DesiredImageCountColumnWidth="{Binding ElementName=Control, Path=DesiredImageCountColumnWidth }"/>
+                <DataTemplate x:DataType="models:ImageFolder">
+                    <local:ImageFolderViewControl Folder="{x:Bind}" SizeChanged="MFImageFolderControl_SizeChanged" DesiredPathColumnWidth="{Binding ElementName=Control, Path=DesiredPathColumnWidth}" DesiredImageCountColumnWidth="{Binding ElementName=Control, Path=DesiredImageCountColumnWidth }"/>
                 </DataTemplate>
             </ListView.ItemTemplate>
             <ListView.ItemsPanel>
@@ -48,7 +49,7 @@
                 </ItemsPanelTemplate>
             </ListView.ItemsPanel>
             <ListView.ItemContainerStyle>
-                <Style x:Key="DefaultListViewItemStyle" TargetType="ListViewItem">
+                <Style TargetType="ListViewItem">
                     <Setter Property="FontFamily" Value="{ThemeResource ContentControlThemeFontFamily}" />
                     <Setter Property="FontSize" Value="{ThemeResource ControlContentThemeFontSize}" />
                     <Setter Property="Background" Value="{ThemeResource ListViewItemBackground}" />
