+++ QuickDrawWindows/Views/ImageFolderListViewControl2.xaml.cs
@@ -1,0 +1,49 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices.WindowsRuntime;
+using Windows.Foundation;
+using Windows.Foundation.Collections;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Controls.Primitives;
+using Microsoft.UI.Xaml.Data;
+using Microsoft.UI.Xaml.Input;
+using Microsoft.UI.Xaml.Media;
+using Microsoft.UI.Xaml.Navigation;
+using QuickDraw.Core.Models;
+using System.Collections.ObjectModel;
+using System.ComponentModel;
+using System.Runtime.InteropServices;
+using CommunityToolkit.WinUI;
+using System.Threading.Tasks;
+using System.Collections.Specialized;
+using QuickDraw.Utilities;
+using System.Diagnostics;
+using WinRT;
+using System.Collections.Concurrent;
+using Windows.System;
+using System.Reflection;
+
+// To learn more about WinUI, the WinUI project structure,
+// and more about our project templates, see: http://aka.ms/winui-project-info.
+
+namespace QuickDraw.Views;
+public sealed partial class MFImageFolderListViewControl2 : UserControl, INotifyPropertyChanged
+{
+    public MFObservableCollection<ImageFolder>? ImageFolderCollection = null;
+    public event PropertyChangedEventHandler? PropertyChanged;
+
+    public MFImageFolderListViewControl2()
+    {
+        this.InitializeComponent();
+
+/*            var settings = (App.Current as App)?.Settings;
+
+        if (settings != null)
+        {
+            ImageFolderCollection = [.. settings.ImageFolderList.ImageFolders];
+        }*/
+    }
+}
