+++ QuickDrawWindows/Views/ImageFolderViewControl.xaml
@@ -1,11 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <UserControl
-    x:Class="QuickDraw.MFImageFolderView"
+    x:Class="QuickDraw.Views.ImageFolderViewControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:QuickDraw"
     xmlns:utilities="using:QuickDraw.Utilities"
-    
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
