+++ QuickDrawWindows/Views/MainPage.xaml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Page
-    x:Class="QuickDraw.MainPage"
+    x:Class="QuickDraw.Views.MainPage"
     x:Name="MainPageElement"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:QuickDraw"
+    xmlns:local="using:QuickDraw.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:slider="using:Syncfusion.UI.Xaml.Sliders"
@@ -38,12 +38,14 @@
         <Grid Padding="24,8,24,24" x:Name="MainGrid" Grid.Row="1" >
             <Grid.RowDefinitions>
                 <RowDefinition Height="*" />
+                <RowDefinition Height="*" />
                 <RowDefinition Height="Auto" />
             </Grid.RowDefinitions>
 
-            <local:MFImageFolderListView x:Name="ImageFolderListView" Grid.Row="0"/>
+            <local:ImageFolderListViewControl x:Name="ImageFolderListView" Grid.Row="0"/>
+            <local:MFImageFolderListViewControl2 x:Name="ImageFolderListView2" Grid.Row="1"/>
 
-            <Grid Grid.Row="1" Margin="0,20,0,0">
+            <Grid Grid.Row="2" Margin="0,20,0,0">
                 <Grid.ColumnDefinitions>
                     <ColumnDefinition Width="*" />
                     <ColumnDefinition Width="Auto" />
